feat: Rebuilding the customization screen: Displaying the customization types.
